[MoE Refactor] Combine MoERunnerBase + DefaultMoERunner
